chore(parquet): move ParquetReader from mod.rs to reader.rs
